refacto(folios) refactor show child folio
