+++ app/validators/folio_validator.ts
@@ -3,7 +3,7 @@ import vine from '@vinejs/vine'
 /**
  * Validates the folio collect action
  */
-export const showValidator = vine.compile(
+export const showChildFolioValidator = vine.compile(
   vine.object({
     params: vine.object({
       id: vine.string(),
