+++ start/routes.ts
@@ -33,7 +33,7 @@ router
             router.get('/', [ChildFoliosController, 'index']).use(middleware.auth())
             router.get('/cards', [MainFoliosController, 'cards']).use(middleware.auth())
             router.get('/statistics', [MainFoliosController, 'statistics']).use(middleware.auth())
-            router.get('/:id', [FoliosController, 'show']).use(middleware.auth())
+            router.get('/:id', [ChildFoliosController, 'show']).use(middleware.auth())
             router.get('/:id/cards', [FoliosController, 'childFolioCards']).use(middleware.auth())
             router.post('/', [CardFoliosController, 'store']).use(middleware.auth())
             router.post('/init', [FoliosController, 'init']).use(middleware.auth())
